refactor: separate prep & submission of note data
... allowing for alternate transports, non-polluting import paths
(previously an import of addNote55 pulled in pouchdb-browser)
