+++ packages/vue/src/db/courseAPI.ts
@@ -2,13 +2,12 @@ import pouch from 'pouchdb-browser';
 import { pouchDBincludeCredentialsConfig } from '.';
 import ENV from '../ENVIRONMENT_VARS';
 import { DataShape } from '../base-course/Interfaces/DataShape';
-import { FieldDefinition } from '../base-course/Interfaces/FieldDefinition';
 import { NameSpacer, ShapeDescriptor } from '../courses/NameSpacer';
-import { FieldType } from '../enums/FieldType';
 import { CourseConfig } from '../server/types';
 import { CourseElo, blankCourseElo, toCourseElo } from '../tutor/Elo';
 import { CourseDB, createTag, updateCardElo } from './courseDB';
 import { CardData, DisplayableData, DocType, Tag } from './types';
+import { prepareNote55 } from './prepareNote55';
 
 /**
  *
@@ -20,7 +19,6 @@ import { CardData, DisplayableData, DocType, Tag } from './types';
  * @param uploads optional additional media uploads: img0, img1, ..., aud0, aud1,...
  * @returns
  */
-
 export async function addNote55(
   courseID: string,
   codeCourse: string,
@@ -31,92 +29,13 @@ export async function addNote55(
   uploads?: { [x: string]: PouchDB.Core.FullAttachment }
 ) {
   const db = await getCourseDB(courseID);
+  const payload = prepareNote55(courseID, codeCourse, shape, data, author, tags, uploads);
+  const result = await db.post<DisplayableData>(payload);
 
   const dataShapeId = NameSpacer.getDataShapeString({
     course: codeCourse,
     dataShape: shape.name,
   });
-
-  const attachmentFields = shape.fields
-    .map((field) => {
-      // make a copy, in order NOT to append to the datashape
-      const copy: FieldDefinition = {
-        name: field.name,
-        type: field.type,
-      };
-      return copy;
-    })
-    .filter((field) => {
-      return field.type === FieldType.IMAGE || field.type === FieldType.AUDIO;
-    })
-    .concat([
-      {
-        name: 'autoplayAudio',
-        type: FieldType.AUDIO,
-      },
-    ]);
-
-  for (let i = 1; i < 11; i++) {
-    if (data[`audio-${i}`]) {
-      attachmentFields.push({
-        name: `audio-${i}`,
-        type: FieldType.AUDIO,
-      });
-    }
-
-    if (data[`image-${i}`]) {
-      attachmentFields.push({
-        name: `image-${i}`,
-        type: FieldType.IMAGE,
-      });
-    }
-  }
-  if (data[`audio-11`]) {
-    throw new Error('Too many audio attachments');
-  }
-  if (data[`image-11`]) {
-    throw new Error('Too many image attachments');
-  }
-
-  const attachments: { [index: string]: PouchDB.Core.FullAttachment } = {};
-  const payload: DisplayableData = {
-    course: courseID,
-    data: [],
-    docType: DocType.DISPLAYABLE_DATA,
-    id_datashape: dataShapeId,
-  };
-
-  if (author) {
-    payload.author = author;
-  }
-
-  attachmentFields.forEach((attField) => {
-    attachments[attField.name] = data[attField.name];
-  });
-
-  //
-  if (uploads) {
-    Object.keys(uploads).forEach((k) => {
-      attachments[k] = uploads[k];
-    });
-  }
-
-  if (attachmentFields.length !== 0 || (uploads && Object.keys(uploads).length)) {
-    payload._attachments = attachments;
-  }
-
-  shape.fields
-    .filter((field) => {
-      return field.type !== FieldType.IMAGE && field.type !== FieldType.AUDIO;
-    })
-    .forEach((field) => {
-      payload.data.push({
-        name: field.name,
-        data: data[field.name],
-      });
-    });
-
-  const result = await db.post<DisplayableData>(payload);
 
   if (result.ok) {
     // create cards
